changed weight of cbes process

+++ Models/CbesProcess.cs
@@ -9,7 +9,7 @@ public partial class CbesProcess
 
     public string? Name { get; set; }
 
-    public int? Weight { get; set; }
+    public decimal? Weight { get; set; }
 
     public DateTime? CreateDate { get; set; }
